refactor(ci): split ref-issues-fetch.mjs out of issue-refs.mjs (module-size gate)
Finding B's --shared-state-file addition pushed issue-refs.mjs to 413 lines
with no scripts/module-size-allowlist.txt row, failing check-module-size.mjs.
Per house rule (AGENTS.md / task instructions), split rather than allowlist:
issue-refs.mjs keeps everything that reads a PR body's own text; the
GraphQL-fetching functions (buildRefsQuery, mapRefsPayload, fetchRefsPayload,
buildRefIssues, fetchRefIssuesIfNeeded) move to the new sibling
scripts/lib/ref-issues-fetch.mjs, with their tests moved to
scripts/lib/ref-issues-fetch.test.mjs. issue-refs.mjs re-exports the two
functions check-issue-queue.mjs needs so that file's own import line, and
line count, is unchanged.

check-module-size.mjs: OK (0 new over 400).
check-test-wiring.mjs: OK (105 scripts/ test files, +1).
